requirements.txt
aiohttp
python-dateutil
yfinance
python-dotenv
pytz
pymongo
apscheduler
pytz
requests
